+++ frameworks/libxgame/src/olua/olua.c
@@ -408,7 +408,7 @@ OLUA_API size_t olua_push_objpool(lua_State *L)
 OLUA_API void olua_pop_objpool(lua_State *L, size_t position)
 {
     if (olua_likely(registry_rawgetp(L, OLUA_POOLTABLE) == LUA_TTABLE)) {
-        size_t len = lua_rawlen(L, -1);
+        size_t len = (size_t)lua_rawlen(L, -1);
         olua_assert(position < len);
         aux_getvmstatus(L)->poolsize = position;
         for (size_t i = position + 1; i <= len; i++) {
